Frontend/backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
pillow==10.0.1
numpy==1.26.0
geopandas==0.14.0
rasterio==1.3.8
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.2
opencv-python==4.8.1.78